Fixed bug with deleting directories
Fixes #6

+++ app/src/main/java/org/emunix/unipatcher/WorkerService.java
@@ -169,7 +169,9 @@ public class WorkerService extends IntentService {
             } else {
                 errorMsg = e.getMessage();
             }
-            FileUtils.deleteQuietly(outputFile);
+            if (outputFile.isFile()) {
+                FileUtils.deleteQuietly(outputFile);
+            }
         } finally {
             stopForeground(true);
         }
